backend/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
psycopg2-binary==2.9.1
PyPDF2==1.26.0
pdfminer.six==20201018
spacy==3.1.3
tensorflow==2.6.0
transformers==4.10.3
celery==5.1.2
pytest==6.2.5
pytest-django==4.4.0
gunicorn==20.1.0
django-cors-headers==3.10.0
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.11.1
boto3==1.18.33
redis==3.5.3
python-decouple==3.4
dj-database-url==0.5.0
whitenoise==5.3.0
django-celery-results==2.0.1
django-celery-beat==2.2.1
Pillow==8.3.2
python-multipart==0.0.5
python-dotenv==0.19.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
scipy==1.7.1
joblib==1.0.1
